Android 17.0.0 release 1

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCajHa3wAKCRDorT+BmrEO
# eKQjAJ9ub9KwPnK4LC/cL8O4Gw8f97wGgQCeOG3McwcjN4maILaVMJ4LN7/mTF0=
# =e1Dp
# -----END PGP SIGNATURE-----
# gpg: Signature made Wed 17 Jun 2026 02:23:11 IDT
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Good signature from "The Android Open Source Project <initial-contribution@android.com>" [unknown]
# gpg: WARNING: This key is not certified with a trusted signature!
# gpg:          There is no indication that the signature belongs to the owner.
# Primary key fingerprint: 4340 D135 70EF 945E 8381  0964 E8AD 3F81 9AB1 0E78

* tag 'android-17.0.0_r1': (57 commits)
  libmeminfo: elf_alignment_test: Use static list for system partitions
  elf_alignment_test: Fix vendor partitions check
  elf_alignment_test: Enforce vendor RELRO alignment from 202604
  libmeminfo: Add support for reading /proc/slabinfo
  Add AndroidBitmapStats to ExtractAndroidHeapStats
  libmeminfo: Add missing test coverage for multiple public APIs
  libmeminfo: Add unit test for MapsWithPageIdle in ProcMemInfo
  Add APK staging support to iter.cpp.
  aac APK/ZIP staging logic replaced with ziparchive_stager.
  Add ziparchive_stager.
  Remove unnecessary windows lib.
  Add support for libelfutils_test to run on Windows.
  Fix 'procrank' tool reporting.
  Add unit test for libelfutils.
  Simplify aac filesystem paths.
  Add a unit test for aac.
  libelfutils: Fix SHT_NOBITS and improve parsing
  Fix swappable_pss calculation
  libmemevents: Increase BPF initialization timeout in test
  libdmabufinfo: Add BPF dmabuf iterator test to presubmit
  ...

+++ libsmapinfo/include/processrecord.h
@@ -36,7 +36,7 @@ class ProcessRecord final {
                   bool get_cmdline, bool get_oomadj, std::ostream& err);
 
     bool valid() const;
-    void CalculateSwap(const std::vector<uint16_t>& swap_offset_array,
+    void CalculateSwap(const std::vector<unsigned int>& swap_offset_array,
                        float zram_compression_ratio);
 
     // Getters
